Moved api test file
